+++ dbt/clients/gcloud.py
@@ -1,7 +1,6 @@
-import subprocess
-
 from dbt.logger import GLOBAL_LOGGER as logger
 import dbt.exceptions
+from dbt.clients.system import run_cmd
 
 NOT_INSTALLED_MSG = """
 dbt requires the gcloud SDK to be installed to authenticate with BigQuery.
@@ -9,21 +8,6 @@ Please download and install the SDK, or use a Service Account instead.
 
 https://cloud.google.com/sdk/
 """
-
-
-def log_cmd(cmd):
-    logger.debug('Executing "{}"'.format(' '.join(cmd)))
-
-
-def run_cmd(cwd, cmd):
-    log_cmd(cmd)
-    proc = subprocess.Popen(
-        cmd,
-        cwd=cwd,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE)
-
-    return proc.communicate()
 
 
 def gcloud_installed():
