+++ dbt/clients/git.py
@@ -1,21 +1,5 @@
-import subprocess
-
+from dbt.clients.system import run_cmd
 from dbt.logger import GLOBAL_LOGGER as logger
-
-
-def log_cmd(cmd):
-    logger.debug('Executing "{}"'.format(' '.join(cmd)))
-
-
-def run_cmd(cwd, cmd):
-    log_cmd(cmd)
-    proc = subprocess.Popen(
-        cmd,
-        cwd=cwd,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE)
-
-    return proc.communicate()
 
 
 def clone(repo, cwd, dirname=None):
